refactor(main): move hasNegativeDimension from core to activity-player
Only used by the main app, no need for it to live in core.
