--- packages/core/src/player/has-negative-dimension.ts
@@ -1,3 +1,0 @@
-export function hasNegativeDimension(dimensions: Record<string, number>): boolean {
-  return Object.values(dimensions).some((value) => value < 0);
-}
